split the files to h and cpp. Fixed the logic. There is some more things to do but in general it should work.

+++ includes/types/position.h
@@ -3,16 +3,16 @@
 #include <utility>
 #include <memory>
 
-using Position = std::pair<int, int>;
+using Position = std::pair<size_t, size_t>;
 
 namespace std
 {
     template <>
     struct hash<Position>
     {
-        size_t operator()(const Position &pos) const
+        size_t operator()(const Position& pos) const
         {
-            return hash<int>()(pos.first) ^ (hash<int>()(pos.second) << 1);
+            return hash<size_t>()(pos.first) ^ (hash<size_t>()(pos.second) << 1);
         }
     };
 }
